Add an option to remove empty SUFFIXED collections when empty

+++ casbah/src/main/scala/akka/contrib/persistence/mongodb/CasbahPersistenceJournaller.scala
@@ -105,7 +105,7 @@ class CasbahPersistenceJournaller(driver: CasbahMongoDriver) extends MongoPersis
     journal.update(query, update, upsert = false, multi = true, writeConcern)
     maxSn.foreach(setMaxSequenceMetadata(persistenceId, _))
     journal.remove(clearEmptyDocumentsQuery(persistenceId), writeConcern)
-    if (driver.useSuffixedCollectionNames && journal.count() == 0)
+    if (driver.useSuffixedCollectionNames && driver.suffixDropEmpty && journal.count() == 0)
       journal.dropCollection()
     ()
   }
